Update README.md
This repository contains the code for a Deepfake Audio Detection project, designed to identify manipulated audio using machine learning techniques. The project includes scripts for preprocessing audio data, training a model, and detecting fake audio. It is designed to be used with any dataset, and the model achieved an accuracy of up to 80% when tested with my own dataset.